Add Exercise15 and update project metadata
This commit introduces Exercise15, which includes a method to sum elements in an array of integers. The `Program.cs` file has been updated to include a case for running Exercise15. DocumentLayout files have been modified to reflect changes in the document structure and state. The AssemblyInfo files have been updated with a new informational version. Several binary files, including DLLs and executables, have been rebuilt, and the source link JSON file now points to a new commit in the GitHub repository.